Sanitize ValueRangeManager
