+++ core/src/test/java/ai/timefold/solver/core/impl/domain/solution/descriptor/SolutionDescriptorTest.java
@@ -5,24 +5,14 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
 import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-import java.util.stream.IntStream;
 
 import ai.timefold.solver.core.api.score.buildin.simple.SimpleScore;
 import ai.timefold.solver.core.impl.score.buildin.SimpleScoreDefinition;
-import ai.timefold.solver.core.impl.score.director.ValueRangeManager;
-import ai.timefold.solver.core.impl.util.MathUtils;
 import ai.timefold.solver.core.testdomain.TestdataEntity;
 import ai.timefold.solver.core.testdomain.TestdataObject;
-import ai.timefold.solver.core.testdomain.TestdataSolution;
 import ai.timefold.solver.core.testdomain.TestdataValue;
-import ai.timefold.solver.core.testdomain.chained.TestdataChainedAnchor;
-import ai.timefold.solver.core.testdomain.chained.TestdataChainedEntity;
-import ai.timefold.solver.core.testdomain.chained.TestdataChainedSolution;
 import ai.timefold.solver.core.testdomain.collection.TestdataArrayBasedSolution;
 import ai.timefold.solver.core.testdomain.collection.TestdataSetBasedSolution;
 import ai.timefold.solver.core.testdomain.declarative.missing.TestdataDeclarativeMissingSupplierSolution;
@@ -41,11 +31,6 @@ import ai.timefold.solver.core.testdomain.invalid.entityannotatedasproblemfact.T
 import ai.timefold.solver.core.testdomain.invalid.multivar.TestdataInvalidMultiVarSolution;
 import ai.timefold.solver.core.testdomain.invalid.nosolution.TestdataNoSolution;
 import ai.timefold.solver.core.testdomain.invalid.variablemap.TestdataMapConfigurationSolution;
-import ai.timefold.solver.core.testdomain.list.TestdataListSolution;
-import ai.timefold.solver.core.testdomain.list.unassignedvar.TestdataAllowsUnassignedValuesListSolution;
-import ai.timefold.solver.core.testdomain.list.valuerange.TestdataListEntityProvidingEntity;
-import ai.timefold.solver.core.testdomain.list.valuerange.TestdataListEntityProvidingSolution;
-import ai.timefold.solver.core.testdomain.list.valuerange.TestdataListEntityProvidingValue;
 import ai.timefold.solver.core.testdomain.reflect.generic.TestdataGenericEntity;
 import ai.timefold.solver.core.testdomain.reflect.generic.TestdataGenericSolution;
 import ai.timefold.solver.core.testdomain.solutionproperties.TestdataNoProblemFactPropertySolution;
@@ -66,16 +51,9 @@ import ai.timefold.solver.core.testdomain.solutionproperties.invalid.TestdataPro
 import ai.timefold.solver.core.testdomain.solutionproperties.invalid.TestdataProblemFactIsPlanningEntityCollectionPropertySolution;
 import ai.timefold.solver.core.testdomain.solutionproperties.invalid.TestdataUnknownFactTypeSolution;
 import ai.timefold.solver.core.testdomain.solutionproperties.invalid.TestdataUnsupportedWildcardSolution;
-import ai.timefold.solver.core.testdomain.unassignedvar.TestdataAllowsUnassignedSolution;
-import ai.timefold.solver.core.testdomain.valuerange.TestdataValueRangeEntity;
-import ai.timefold.solver.core.testdomain.valuerange.TestdataValueRangeSolution;
-import ai.timefold.solver.core.testdomain.valuerange.entityproviding.unassignedvar.TestdataAllowsUnassignedEntityProvidingEntity;
-import ai.timefold.solver.core.testdomain.valuerange.entityproviding.unassignedvar.TestdataAllowsUnassignedEntityProvidingSolution;
 import ai.timefold.solver.core.testutil.CodeAssertableArrayList;
 import ai.timefold.solver.core.testutil.PlannerTestUtils;
 
-import org.assertj.core.data.Percentage;
-import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 
 class SolutionDescriptorTest {
@@ -361,312 +339,6 @@ class SolutionDescriptorTest {
 
         assertAllCodesOfCollection(solutionDescriptor.getAllEntitiesAndProblemFacts(solution), "otherE1", "f1", "p1", "e1",
                 "e2");
-    }
-
-    @Test
-    void countEntities() {
-        var valueCount = 10;
-        var entityCount = 3;
-        var solution = TestdataListSolution.generateInitializedSolution(valueCount, entityCount);
-        var solutionDescriptor = TestdataListSolution.buildSolutionDescriptor();
-        var valueRangeManager = new ValueRangeManager<TestdataListSolution>(solutionDescriptor);
-        var initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.genuineEntityCount()).isEqualTo(entityCount);
-        assertThat(initializationStats.shadowEntityCount()).isEqualTo(valueCount);
-    }
-
-    @Test
-    void countUninitializedVariables() {
-        var valueCount = 10;
-        var entityCount = 3;
-        var solution = TestdataSolution.generateSolution(valueCount, entityCount);
-        var solutionDescriptor = TestdataSolution.buildSolutionDescriptor();
-
-        var valueRangeManager = new ValueRangeManager<TestdataSolution>(solutionDescriptor);
-        var initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.uninitializedVariableCount()).isZero();
-
-        solution.getEntityList().get(0).setValue(null);
-        initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.uninitializedVariableCount()).isOne();
-
-        solution.getEntityList().forEach(entity -> entity.setValue(null));
-        initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.uninitializedVariableCount()).isEqualTo(entityCount);
-    }
-
-    @Test
-    void countUnassignedValues() {
-        var valueCount = 10;
-        var entityCount = 3;
-        var solution = TestdataListSolution.generateInitializedSolution(valueCount, entityCount);
-        var solutionDescriptor = TestdataListSolution.buildSolutionDescriptor();
-        var valueRangeManager = new ValueRangeManager<TestdataListSolution>(solutionDescriptor);
-
-        var initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.unassignedValueCount()).isZero();
-
-        var valueList = solution.getEntityList().get(0).getValueList();
-        var unassignedValueCount = valueList.size();
-        assertThat(valueList).hasSizeGreaterThan(10 / 3);
-        valueList.forEach(value -> {
-            value.setEntity(null);
-            value.setIndex(null);
-        });
-        valueList.clear();
-
-        initializationStats = solutionDescriptor.computeInitializationStatistics(solution, valueRangeManager);
-        assertThat(initializationStats.unassignedValueCount()).isEqualTo(unassignedValueCount);
-    }
-
-    @Test
-    void problemScaleBasic() {
-        var valueCount = 10;
-        var entityCount = 20;
-        var solutionDescriptor = TestdataSolution.buildSolutionDescriptor();
-        var solution = TestdataSolution.generateSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isEqualTo(20.0);
-        });
-    }
-
-    @Test
-    void emptyProblemScale() {
-        var valueCount = 27;
-        var entityCount = 27;
-        var solutionDescriptor = TestdataSolution.buildSolutionDescriptor();
-        var solution = TestdataSolution.generateSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        solution.getValueList().clear();
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThatThrownBy(() -> solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager))
-                    .isInstanceOf(IllegalStateException.class);
-            softly.assertThatThrownBy(() -> solutionDescriptor.getApproximateValueCount(solution, valueRangeManager))
-                    .isInstanceOf(IllegalStateException.class);
-            softly.assertThatThrownBy(() -> solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isInstanceOf(IllegalStateException.class);
-        });
-    }
-
-    @Test
-    void emptyProblemScaleAllowsUnassigned() {
-        var valueCount = 27;
-        var entityCount = 27;
-        var solutionDescriptor = TestdataAllowsUnassignedSolution.buildSolutionDescriptor();
-        var solution = TestdataAllowsUnassignedSolution.generateSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        solution.getValueList().clear();
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(1);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(1);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isEqualTo(0);
-        });
-    }
-
-    @Test
-    void problemScaleMultipleValueRanges() {
-        var solutionDescriptor = TestdataValueRangeSolution.buildSolutionDescriptor();
-        var solution = new TestdataValueRangeSolution("Solution");
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        solution.setEntityList(List.of(new TestdataValueRangeEntity("A")));
-        final var entityCount = 1L;
-        final var valueCount = 3L;
-        final var variableCount = 8L;
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount * variableCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(3L);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager))
-                    .isEqualTo(variableCount * valueCount);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(Math.log10(Math.pow(valueCount, variableCount)), Percentage.withPercentage(1.0));
-        });
-    }
-
-    @Test
-    void basicVariableProblemScaleEntityProvidingValueRange() {
-        var solutionDescriptor = TestdataAllowsUnassignedEntityProvidingSolution.buildSolutionDescriptor();
-        var solution = new TestdataAllowsUnassignedEntityProvidingSolution("Solution");
-        var v1 = new TestdataValue("1");
-        var v2 = new TestdataValue("2");
-        solution.setEntityList(List.of(
-                new TestdataAllowsUnassignedEntityProvidingEntity("A",
-                        List.of(v1, v2)),
-                new TestdataAllowsUnassignedEntityProvidingEntity("B",
-                        List.of(v1, v2, new TestdataValue("3")))));
-
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        valueRangeManager.reset(solution);
-
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(2L);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(2L);
-
-            // Add 1 to the value range sizes, since the value range allows unassigned
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(4L);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(3L + 4L);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(Math.log10(3 * 4), Percentage.withPercentage(1.0));
-        });
-    }
-
-    @Disabled("Temporarily disabled")
-    @Test
-    void listVariableProblemScaleEntityProvidingValueRange() {
-        var solutionDescriptor = TestdataListEntityProvidingSolution.buildSolutionDescriptor();
-        var solution = new TestdataListEntityProvidingSolution();
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        var v1 = new TestdataListEntityProvidingValue("1");
-        var v2 = new TestdataListEntityProvidingValue("2");
-        solution.setEntityList(List.of(
-                new TestdataListEntityProvidingEntity("e1", List.of(v1, v2)),
-                new TestdataListEntityProvidingEntity("e2", List.of(v1, v2, new TestdataListEntityProvidingValue("3")))));
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(2L);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(2L);
-
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(3L);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(2L + 3L);
-        });
-    }
-
-    @Test
-    void problemScaleSingleEntityProvidingSingleValueRange() {
-        var solutionDescriptor = TestdataAllowsUnassignedEntityProvidingSolution.buildSolutionDescriptor();
-        var solution = new TestdataAllowsUnassignedEntityProvidingSolution("Solution");
-
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        valueRangeManager.reset(solution);
-
-        var v1 = new TestdataValue("1");
-        solution.setEntityList(List.of(
-                new TestdataAllowsUnassignedEntityProvidingEntity("A",
-                        List.of(v1))));
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(1L);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(1L);
-
-            // Add 1 to the value range sizes, since the value range allows unassigned
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(2L);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(2L);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(Math.log10(2), Percentage.withPercentage(1.0));
-        });
-    }
-
-    @Test
-    void problemScaleChained() {
-        var anchorCount = 20;
-        var entityCount = 500;
-        var solutionDescriptor = TestdataChainedSolution.buildSolutionDescriptor();
-        var solution = generateChainedSolution(anchorCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount * 2);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager))
-                    .isEqualTo(entityCount + anchorCount);
-            // 1 unchained value is inside the solution
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager))
-                    .isEqualTo(entityCount + anchorCount + 1);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(MathUtils.getPossibleArrangementsScaledApproximateLog(MathUtils.LOG_PRECISION, 10, 500, 20)
-                            / (double) MathUtils.LOG_PRECISION, Percentage.withPercentage(1.0));
-        });
-    }
-
-    static TestdataChainedSolution generateChainedSolution(int anchorCount, int entityCount) {
-        var solution = new TestdataChainedSolution("test solution");
-        var anchorList = IntStream.range(0, anchorCount)
-                .mapToObj(Integer::toString)
-                .map(TestdataChainedAnchor::new).toList();
-        solution.setChainedAnchorList(anchorList);
-        var entityList = IntStream.range(0, entityCount)
-                .mapToObj(Integer::toString)
-                .map(TestdataChainedEntity::new).toList();
-        solution.setChainedEntityList(entityList);
-        solution.setUnchainedValueList(Collections.singletonList(new TestdataValue("v")));
-        return solution;
-    }
-
-    @Test
-    void problemScaleList() {
-        var valueCount = 500;
-        var entityCount = 20;
-        var solutionDescriptor = TestdataListSolution.buildSolutionDescriptor();
-        var solution = TestdataListSolution.generateUninitializedSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(MathUtils.getPossibleArrangementsScaledApproximateLog(MathUtils.LOG_PRECISION, 10, 500, 20)
-                            / (double) MathUtils.LOG_PRECISION, Percentage.withPercentage(1.0));
-        });
-    }
-
-    @Test
-    void problemScaleSingleEntityWithAssignedValues() {
-        var valueCount = 1;
-        var entityCount = 1;
-        var solutionDescriptor = TestdataListSolution.buildSolutionDescriptor();
-        var solution = TestdataListSolution.generateUninitializedSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager)).isEqualTo(0.0);
-        });
-    }
-
-    @Test
-    void problemScaleSingleEntityWithUnassignedValues() {
-        var valueCount = 1;
-        var entityCount = 1;
-        var solutionDescriptor = TestdataAllowsUnassignedValuesListSolution.buildSolutionDescriptor();
-        var solution = TestdataAllowsUnassignedValuesListSolution.generateUninitializedSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        assertSoftly(softly -> {
-            softly.assertThat(solutionDescriptor.getGenuineEntityCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getGenuineVariableCount(solution)).isEqualTo(entityCount);
-            softly.assertThat(solutionDescriptor.getMaximumValueRangeSize(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getApproximateValueCount(solution, valueRangeManager)).isEqualTo(valueCount);
-            softly.assertThat(solutionDescriptor.getProblemScale(solution, valueRangeManager))
-                    .isCloseTo(Math.log10(2), Percentage.withPercentage(1.0));
-        });
-    }
-
-    @Test
-    void assertProblemScaleListIsApproximatelyProblemScaleChained() {
-        var valueCount = 500;
-        var entityCount = 20;
-        var solutionDescriptor = TestdataListSolution.buildSolutionDescriptor();
-        var listSolution = TestdataListSolution.generateUninitializedSolution(valueCount, entityCount);
-        var valueRangeManager = new ValueRangeManager<>(solutionDescriptor);
-        var listPowerExponent = solutionDescriptor.getProblemScale(listSolution, valueRangeManager);
-        var solutionDescriptorChained = TestdataChainedSolution.buildSolutionDescriptor();
-        var solutionChained = generateChainedSolution(entityCount, valueCount);
-        var valueRangeManagerChained = new ValueRangeManager<>(solutionDescriptorChained);
-        var chainedPowerExponent = solutionDescriptorChained.getProblemScale(solutionChained, valueRangeManagerChained);
-        // Since they are using different bases in calculation, some difference is expected,
-        // but the numbers should be relatively (i.e. ~1%) close.
-        assertThat(Math.pow(10, listPowerExponent))
-                .isCloseTo(Math.pow(10, chainedPowerExponent), Percentage.withPercentage(1));
     }
 
     @Test
